Add payload, utils files
